# src/Api/Dtos/OrganisationRegistration.cs
using System.Text.Json.Serialization;

namespace Api.Dtos;

public record OrganisationRegistration
{
    [JsonPropertyName("name")]
    public required string Name { get; init; }

    [JsonPropertyName("tradingName")]
    public string? TradingName { get; init; }

    [JsonPropertyName("businessCountry")]
    public BusinessCountry? BusinessCountry { get; init; }

    [JsonPropertyName("companiesHouseNumber")]
    public string? CompaniesHouseNumber { get; init; }

    [JsonPropertyName("address")]
    public required Address Address { get; init; }

    [JsonPropertyName("registration")]
    public required Registration Registration { get; init; }
}